requirements.txt
pytest
selenium==3.141.0
jsonpickle
pony
pymysql
mysql-connector-python-rf
allure-pytest
